// frontend/src/sections/patients/detail/updatePatient.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type UpdatePatientVariables = {
  storeId: string;
  input: {
    id: string;
    code: string;
    code2?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    gender?: "FEMALE" | "MALE" | "TRANSGENDER" | "TRANSGENDER_MALE" | "TRANSGENDER_MALE_HORMONE" | "TRANSGENDER_MALE_SURGICAL" | "TRANSGENDER_FEMALE" | "TRANSGENDER_FEMALE_HORMONE" | "TRANSGENDER_FEMALE_SURGICAL" | "UNKNOWN" | "NON_BINARY" | null;
    dateOfBirth?: string | null;
    address1?: string | null;
    phone?: string | null;
    isDeceased?: boolean | null;
    dateOfDeath?: string | null;
    nextOfKinId?: string | null;
    nextOfKinName?: string | null;
  };
};

export type UpdatePatientResult = {
  updatePatient: ({
  __typename: "PatientNode";
} & {
  id: string;
});
};

export const UpdatePatient = {
  query: "mutation updatePatient($storeId: String!, $input: UpdatePatientInput!) {\n  updatePatient(storeId: $storeId, input: $input) {\n    ... on PatientNode {\n      __typename\n      id\n    }\n  }\n}",
} as TypedDocument<UpdatePatientResult, UpdatePatientVariables>;

export type UpdatePatientCustomFieldsVariables = {
  storeId: string;
  input: {
    id: string;
    customFields: unknown;
  };
};

export type UpdatePatientCustomFieldsResult = {
  updatePatientCustomFields: ({
  __typename: "PatientNode";
} & {
  id: string;
  customFields: unknown | null;
});
};

export const UpdatePatientCustomFields = {
  query: "mutation updatePatientCustomFields($storeId: String!, $input: UpdatePatientCustomFieldsInput!) {\n  updatePatientCustomFields(storeId: $storeId, input: $input) {\n    ... on PatientNode {\n      __typename\n      id\n      customFields\n    }\n  }\n}",
} as TypedDocument<UpdatePatientCustomFieldsResult, UpdatePatientCustomFieldsVariables>;